+++ backend/src/routes/projects.js
@@ -1,0 +1,27 @@
+const controller = require("../controllers/projects");
+const validators = require("../validators/projects");
+const what = "projects";
+
+const routes = [
+  {
+    method: "GET",
+    url: `/${what}`,
+    handler: controller.getAll,
+  },
+  {
+    method: "GET",
+    url: `/${what}/:id`,
+    schema: validators.getOneValidator,
+    handler: controller.getOne,
+  },
+];
+
+const registerRoutes = (fastify, options, done) => {
+  // Ensure all of the routes above get registered.
+  routes.forEach((route) => fastify.route(route));
+  done();
+};
+
+module.exports = {
+  registerRoutes,
+};